--- base/state.py
@@ -1,3 +1,0 @@
-global_state = {
-    "sessions": {}  # Stores session and user information
-}
