+++ poker/view_points.py
@@ -1,10 +1,8 @@
-import time
-import json
-import sqlite3
+# Description: This file contains the code to view the selected points for a session and all active sessions.
 import streamlit as st
 import pandas as pd
 
-from util import delete_session, display_session_data
+from util import delete_session, display_session_data, display_all_session_data, delete_all_sessions
 
 def view_points():
     st.title("View Selected Points")
@@ -25,4 +23,13 @@ def view_points():
         delete_session(session_id_int)
         st.success("Session data deleted successfully!")
         st.rerun()
-# active_sessions()
+
+def all_active_sessions():
+    st.title("Active Sessions")
+    display_all_session_data()
+    
+    # delete_all_sessions()
+    if st.button("Delete All Sessions"):
+        delete_all_sessions()
+        st.success("All sessions deleted successfully!")
+        st.rerun()
