--- run.md
@@ -1,12 +1,0 @@
-
-```bash
-streamlit run src/app.py
-```
-```bash
-streamlit run base/app.py
-```
-
-
-```bash
-streamlit run base/login.py
-```
